Separate analysis io functions from text module
text module is reserved for functions concerning reading data, since we may add
one module per input kind
